requirements.txt
Django==1.6.5
South==0.8.4
dj-database-url==0.3.0
dj-static==0.0.5
django-allauth==0.17.0
django-toolbelt==0.0.1
gunicorn==19.0.0
oauthlib==0.6.3
psycopg2==2.5.3
pystache==0.5.4
python-openid==2.2.5
requests==2.3.0
requests-oauthlib==0.4.1
static==1.0.2
wsgiref==0.1.2
